fixed comments left on pr
